Add files via upload
This Computer Vision are able to detect Bolt in real time with mAP50 score at 88.8%. The computer vision is developed using YOLOv8, which is an algorithm derived from Convolutional Neural Network (CNN), a deep learning model. this is a part of my Final Year Project (FYP) in Universiti Malaya where I need to develop a robot arm for automated fishplate bolt tightening in railway maintenance. Thsi algorithm will then be connected with the end effector of my robot arm as a guidance for the robot arm to automatically tighten the bolt.